Add function to log in browser console
