+++ nativemessaging/__init__.py
@@ -1,28 +1,3 @@
-#!/usr/bin/env python
-import json
-import sys
-import struct
+# -*- coding: utf-8 -*-
 
-
-# from https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Native_messaging
-def get_message():
-    raw_length = sys.stdin.buffer.read(4)
-    if len(raw_length) == 0:
-        sys.exit(0)
-    message_length = struct.unpack("@I", raw_length)[0]
-    message = sys.stdin.buffer.read(message_length).decode("utf-8")
-    return json.loads(message)
-
-
-# Encode a message for transmission, given its content.
-def encode_message(message_content):
-    encoded_content = json.dumps(message_content).encode("utf-8")
-    encoded_length = struct.pack("@I", len(encoded_content))
-    return {"length": encoded_length, "content": encoded_content}
-
-
-# Send an encoded message to stdout.
-def send_message(encoded_message):
-    sys.stdout.buffer.write(encoded_message["length"])
-    sys.stdout.buffer.write(encoded_message["content"])
-    sys.stdout.buffer.flush()
+from nativemessaging import *
